+++ __tests__/containers/CodesplainAppBar.test.jsx
@@ -3,8 +3,6 @@ import cookie from 'react-cookie';
 import { shallow } from 'enzyme';
 import { shallowToJson } from 'enzyme-to-json';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import { createStore } from 'redux';
-import app from '../../src/reducers/app.js'
 
 import { CodesplainAppBar } from '../../src/containers/CodesplainAppBar';
 
@@ -13,13 +11,16 @@ describe('<CodesplainAppBar />', () => {
   const shallowWithContext = (node) => shallow(node, { context: { muiTheme } });
 
   describe('snapshot tests', () => {
+    const userState = {
+      userSnippets: [],
+    };
     afterEach(() => {
       cookie.remove('token');
     });
     it('user is not logged in', () => {
       const wrapper = shallowWithContext(
         <CodesplainAppBar
-          store={createStore(app)}
+          userState={userState}
         />
       );
       expect(shallowToJson(wrapper)).toMatchSnapshot();
@@ -29,7 +30,7 @@ describe('<CodesplainAppBar />', () => {
       cookie.save('token', 'foobar')
       const wrapper = shallowWithContext(
         <CodesplainAppBar
-          store={createStore(app)}
+          userState={userState}
         />
       );
       expect(shallowToJson(wrapper)).toMatchSnapshot();
